Moved editor shared parts to base
