+++ src/components/Editors/Chart/Layout.tsx
@@ -4,7 +4,7 @@ import { useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
 import Columns from '../../Parts/Columns'
 import VerticalTabs from '../../Parts/Tabs/Vertical'
-import EditorHelp from '../../Parts/Editor/Help'
+import EditorHelp from '../Base/Help'
 import Chart from './Sections/Chart/Chart'
 import Channel from './Sections/Channel/Channel'
 import Transform from './Sections/Transform/Transform'
