+++ src/components/Editors/Chart/Sections/Chart/Chart.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import InputField from '../../../../Parts/Fields/Input'
 import SelectField from '../../../../Parts/Fields/Select'
-import EditorSection from '../../../../Parts/Editor/Section'
+import EditorSection from '../../../Base/Section'
 import Columns from '../../../../Parts/Columns'
 import { useStore, selectors } from '../../store'
 import * as settings from '../../settings'
