+++ src/components/Editors/Chart/Sections/Transform/Transform.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
 import Columns from '../../../../Parts/Columns'
-import EditorItem from '../../../../Parts/Editor/Item'
-import EditorList from '../../../../Parts/Editor/List'
-import EditorListItem from '../../../../Parts/Editor/ListItem'
-import EditorSearch from '../../../../Parts/Editor/Search'
+import EditorItem from '../../../Base/Item'
+import EditorList from '../../../Base/List'
+import EditorListItem from '../../../Base/ListItem'
+import EditorSearch from '../../../Base/Search'
 import SelectField from '../../../../Parts/Fields/Select'
 import { useStore, selectors, select } from '../../store'
 import Aggregate from './Aggregate'
