+++ src/components/Editors/Dialect/Sections/Dialect.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import InputField from '../../../Parts/Fields/Input'
 import YesNoField from '../../../Parts/Fields/YesNo'
 import MultilineField from '../../../Parts/Fields/Multiline'
-import EditorSection from '../../../Parts/Editor/Section'
+import EditorSection from '../../Base/Section'
 import Columns from '../../../Parts/Columns'
 import * as settings from '../../../../settings'
 import { useStore } from '../store'
