+++ src/components/Editors/Dialect/Sections/Html.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import Columns from '../../../Parts/Columns'
 import InputField from '../../../Parts/Fields/Input'
-import EditorSection from '../../../Parts/Editor/Section'
+import EditorSection from '../../Base/Section'
 import * as settings from '../../../../settings'
 import { useStore, selectors, select } from '../store'
