+++ src/components/Editors/Package/Layout.tsx
@@ -5,7 +5,7 @@ import Box from '@mui/material/Box'
 import InputAdornment from '@mui/material/InputAdornment'
 import HorizontalTabs from '../../Parts/Tabs/Horizontal'
 import VerticalTabs from '../../Parts/Tabs/Vertical'
-import EditorHelp from '../../Parts/Editor/Help'
+import EditorHelp from '../Base/Help'
 import SelectField from '../../Parts/Fields/Select'
 import Columns from '../../Parts/Columns'
 import Resource from '../Resource'
