+++ src/components/Editors/Package/Sections/Package.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import InputField from '../../../Parts/Fields/Input'
 import DatePickerField from '../../../Parts/Fields/DatePicker'
 import MultilineField from '../../../Parts/Fields/Multiline'
-import EditorSection from '../../../Parts/Editor/Section'
+import EditorSection from '../../Base/Section'
 import Columns from '../../../Parts/Columns'
 import { useStore } from '../store'
 import validator from 'validator'
