+++ src/components/Editors/Resource/Layout.tsx
@@ -3,7 +3,7 @@ import camelCase from 'lodash/camelCase'
 import { useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
 import Columns from '../../Parts/Columns'
-import EditorHelp from '../../Parts/Editor/Help'
+import EditorHelp from '../Base/Help'
 import HorizontalTabs from '../../Parts/Tabs/Horizontal'
 import VerticalTabs from '../../Parts/Tabs/Vertical'
 import Dialect from '../Dialect'
