+++ src/components/Editors/Schema/Sections/Field.tsx
@@ -3,10 +3,10 @@ import partition from 'lodash/partition'
 import Box from '@mui/material/Box'
 import { useTheme } from '@mui/material/styles'
 import Columns from '../../../Parts/Columns'
-import EditorItem from '../../../Parts/Editor/Item'
-import EditorList from '../../../Parts/Editor/List'
-import EditorListItem from '../../../Parts/Editor/ListItem'
-import EditorSearch from '../../../Parts/Editor/Search'
+import EditorItem from '../../Base/Item'
+import EditorList from '../../Base/List'
+import EditorListItem from '../../Base/ListItem'
+import EditorSearch from '../../Base/Search'
 import InputField from '../../../Parts/Fields/Input'
 import YesNoField from '../../../Parts/Fields/YesNo'
 import SelectField from '../../../Parts/Fields/Select'
