+++ src/components/Editors/Schema/Sections/ForeignKey.tsx
@@ -4,10 +4,10 @@ import { useTheme } from '@mui/material/styles'
 import Columns from '../../../Parts/Columns'
 import InputField from '../../../Parts/Fields/Input'
 import SelectField from '../../../Parts/Fields/Select'
-import EditorItem from '../../../Parts/Editor/Item'
-import EditorList from '../../../Parts/Editor/List'
-import EditorListItem from '../../../Parts/Editor/ListItem'
-import EditorSearch from '../../../Parts/Editor/Search'
+import EditorItem from '../../Base/Item'
+import EditorList from '../../Base/List'
+import EditorListItem from '../../Base/ListItem'
+import EditorSearch from '../../Base/Search'
 import { useStore, selectors, select } from '../store'
 import ScrollBox from '../../../Parts/ScrollBox'
