+++ src/components/Editors/Schema/Sections/Schema.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import InputField from '../../../Parts/Fields/Input'
 import MultiselectField from '../../../Parts/Fields/Multiselect'
 import MultilineField from '../../../Parts/Fields/Multiline'
-import EditorSection from '../../../Parts/Editor/Section'
+import EditorSection from '../../Base/Section'
 import Columns from '../../../Parts/Columns'
 import { useStore, selectors } from '../store'
 import validator from 'validator'
