+++ src/components/Editors/View/Layout.tsx
@@ -3,7 +3,7 @@ import { useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
 import Columns from '../../Parts/Columns'
 import Query from './Sections/Query'
-import EditorHelp from '../../Parts/Editor/Help'
+import EditorHelp from '../Base/Help'
 import { useStore } from './store'
 
 export default function Layout() {
