+++ src/components/Editors/View/Sections/Query.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
-import EditorSection from '../../../Parts/Editor/Section'
-import TextEditor from '../../../Editors/Text'
+import TextEditor from '../../Text'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import ColumnTree from '../../../Parts/Trees/Column'
-import EditorSearch from '../../../Parts/Editor/Search'
+import EditorSection from '../../Base/Section'
+import EditorSearch from '../../Base/Search'
 import ScrollBox from '../../../Parts/ScrollBox'
 import Columns from '../../../Parts/Columns'
 import { useTheme } from '@mui/material/styles'
