+++ src/components/Editors/View/store.ts
@@ -5,7 +5,7 @@ import cloneDeep from 'lodash/cloneDeep'
 import { assert } from 'ts-essentials'
 import { createStore } from 'zustand/vanilla'
 import { createSelector } from 'reselect'
-import { ITextEditor } from '../../Editors/Text'
+import { ITextEditor } from '../Text'
 import { ViewProps } from './index'
 import * as settings from '../../../settings'
 import * as helpers from '../../../helpers'
